refactor: criado pacote Dominio
