+++ src/Dominio/Cliente.java
@@ -1,3 +1,4 @@
+package Dominio;
 
 public class Cliente {
